+++ shared/src/main/scala/com/kodekutters/stix/Objects.scala
@@ -204,20 +204,7 @@ object MarkingObject {
   * and Marking Definitions.
   */
 case class GranularMarking(selectors: List[String], marking_ref: Option[String] = None, lang: Option[String] = None) {
-
   val `type` = GranularMarking.`type`
-
-  override def toString = {
-    val sb1 = new StringBuilder()
-    for (s <- selectors) sb1.append(s + ";")
-    // remove the last ";"
-    val sb = new StringBuilder(sb1.toString().reverse.substring(1).reverse)
-    sb.append(marking_ref.getOrElse(""))
-    sb.append(",")
-    sb.append(lang.getOrElse(""))
-    sb.toString()
-  }
-
 }
 
 object GranularMarking {
@@ -229,10 +216,11 @@ object GranularMarking {
 /**
   * External references are used to describe pointers to information represented outside of STIX.
   */
-case class ExternalReference(source_name: String, description: Option[String] = None, url: Option[String] = None, external_id: Option[String] = None) {
+case class ExternalReference(source_name: String, description: Option[String] = None,
+                             external_id: Option[String] = None,
+                             url: Option[String] = None,
+                             hashes: Option[Map[String, String]] = None) {
   val `type` = ExternalReference.`type`
-
-  override def toString = source_name + "," + description.getOrElse("") + "," + url.getOrElse("") + "," + external_id.getOrElse("")
 }
 
 object ExternalReference {